Adds working parser with replacements and variables
This feels kind of cool! You can define set characters or adjectives
or anything in the story liners, and if they exist, they'll be used.
If they don't exist, they'll be generated, and will propagate through
the rest of the story as it's created.

Next is importing from a file instead of reading from a predefined
string, and expanding the rules. Also, improve the error handling.